# openApi.yml
openapi: "3.0.2"
info:
  title: "api documention"
  description: "this is document page of our apis"
  version: "2.0.0"
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /developer/hash-string/{string}:
    get:
      summary: hash one string
      tags: [Developers]
      description: hash one string with bcrypt
      parameters:
        - in: path
          name: string
          schema:
            type: string
          required: true
      responses:
        200:
          description: success
        400:
          description: bad request
        500:
          description: internal server error
  /developer/random-number:
    get:
      summary: get a random number
      tags: [Developers]
      description: get a random number
      responses:
        200:
          description: success
        400:
          description: bad request
        500:
          description: internal server error
...